+++ tests/Feature/ListenersTest.php
@@ -3,11 +3,11 @@
 namespace Sfneal\Tracking\Tests\Feature;
 
 use Illuminate\Support\Facades\Event;
+use Sfneal\Testing\Utils\Traits\EventFaker;
 use Sfneal\Tracking\Events\TrackActivityEvent;
 use Sfneal\Tracking\Events\TrackTrafficEvent;
 use Sfneal\Tracking\Listeners\TrackActivityListener;
 use Sfneal\Tracking\Listeners\TrackTrafficListener;
-use Sfneal\Tracking\Tests\EventFaker;
 use Sfneal\Tracking\Tests\TestCase;
 
 class ListenersTest extends TestCase