+++ tests/Unit/TrackActionTest.php
@@ -3,9 +3,9 @@
 namespace Sfneal\Tracking\Tests\Unit;
 
 use Sfneal\Testing\Models\People;
+use Sfneal\Testing\Utils\Interfaces\CrudModelTest;
 use Sfneal\Tracking\Jobs\TrackActionJob;
 use Sfneal\Tracking\Models\TrackAction;
-use Sfneal\Tracking\Tests\CrudModelTest;
 use Sfneal\Tracking\Tests\TestCase;
 
 class TrackActionTest extends TestCase implements CrudModelTest