+++ tests/Unit/TrackTrafficTest.php
@@ -2,11 +2,11 @@
 
 namespace Sfneal\Tracking\Tests\Unit;
 
+use Sfneal\Testing\Utils\Interfaces\CrudModelTest;
+use Sfneal\Testing\Utils\Traits\CreateRequest;
 use Sfneal\Tracking\Actions\TrackTrafficAction;
 use Sfneal\Tracking\Events\TrackTrafficEvent;
 use Sfneal\Tracking\Models\TrackTraffic;
-use Sfneal\Tracking\Tests\CreateRequest;
-use Sfneal\Tracking\Tests\CrudModelTest;
 use Sfneal\Tracking\Tests\TestCase;
 
 class TrackTrafficTest extends TestCase implements CrudModelTest